# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SquadPlayer {
  name: string;
  /**
     * @minimum 15
     * @maximum 45
     */
  age: number;
  /**
     * @minimum 1
     * @maximum 99
     */
  rating: number;
  nationality: string;
  position: string;
  number: string;
}

export interface SquadGenerationResponse {
  players: SquadPlayer[];
}

export interface SquadFromTextRequest {
  /** @minLength 1 */
  rawText: string;
  sport?: string;
  clubName?: string;
}

export type SquadFromCriteriaRequestCriteria = {
  sport: string;
  clubName?: string;
  /**
     * @minimum 1
     * @maximum 40
     */
  count: number;
  /**
     * @minimum 15
     * @maximum 45
     */
  ageMin?: number;
  /**
     * @minimum 15
     * @maximum 45
     */
  ageMax?: number;
  /**
     * @minimum 1
     * @maximum 99
     */
  ratingMin?: number;
  /**
     * @minimum 1
     * @maximum 99
     */
  ratingMax?: number;
  nationality?: string;
  positions?: string;
};

export interface SquadFromCriteriaRequest {
  criteria: SquadFromCriteriaRequestCriteria;
}

export interface BestXIPlayer {
  id: string;
  name: string;
  position: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  rating: number;
  number: string;
}

export interface BestXIRequest {
  sport: string;
  formation: string;
  /** @maxItems 40 */
  players: BestXIPlayer[];
}

export interface BestXIResponse {
  playerIds: string[];
}

export interface SharedRosterPlayer {
  id: string;
  name: string;
  age: number;
  rating: number;
  nationality: string;
  position: string;
  number: string;
}

export type SharedClubCategory = typeof SharedClubCategory[keyof typeof SharedClubCategory];


export const SharedClubCategory = {
  club: 'club',
  selection: 'selection',
} as const;

export interface SharedClub {
  id?: string;
  name: string;
  logo?: string;
  jerseyColor: string;
  secondaryColor: string;
  accentColor: string;
  numberColor: string;
  jerseyStyle: string;
  backgroundId: string;
  category: SharedClubCategory;
  roster?: SharedRosterPlayer[];
}

export interface SharedClubResponse {
  club: SharedClub;
}

export interface SharedClubsResponse {
  clubs: SharedClub[];
}

export interface SharedPlayer { [key: string]: unknown }

export type SharedCompositionSport = typeof SharedCompositionSport[keyof typeof SharedCompositionSport];


export const SharedCompositionSport = {
  football: 'football',
  hockey: 'hockey',
} as const;

export interface SharedComposition {
  id?: string;
  name: string;
  savedAt?: number;
  sport: SharedCompositionSport;
  formation: string;
  title: string;
  players: SharedPlayer[];
  bench: SharedPlayer[];
  jerseyColor: string;
  secondaryColor: string;
  accentColor: string;
  numberColor?: string;
  jerseyStyle: string;
  backgroundId: string;
  showBench: boolean;
  showDetails: boolean;
}

export interface SharedCompositionResponse {
  composition: SharedComposition;
}

export interface SharedCompositionsResponse {
  compositions: SharedComposition[];
}