# lib/api-zod/src/generated/types/bestXIRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BestXIPlayer } from './bestXIPlayer';

export interface BestXIRequest {
  sport: string;
  formation: string;
  /** @maxItems 40 */
  players: BestXIPlayer[];
}