# lib/api-zod/src/generated/types/sharedClub.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SharedClubCategory } from './sharedClubCategory';
import type { SharedRosterPlayer } from './sharedRosterPlayer';

export interface SharedClub {
  id?: string;
  name: string;
  logo?: string;
  jerseyColor: string;
  secondaryColor: string;
  accentColor: string;
  numberColor: string;
  jerseyStyle: string;
  backgroundId: string;
  category: SharedClubCategory;
  roster?: SharedRosterPlayer[];
}