# lib/api-zod/src/generated/types/sharedClubsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SharedClub } from './sharedClub';

export interface SharedClubsResponse {
  clubs: SharedClub[];
}