# lib/api-zod/src/generated/types/sharedComposition.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SharedCompositionSport } from './sharedCompositionSport';
import type { SharedPlayer } from './sharedPlayer';

export interface SharedComposition {
  id?: string;
  name: string;
  savedAt?: number;
  sport: SharedCompositionSport;
  formation: string;
  title: string;
  players: SharedPlayer[];
  bench: SharedPlayer[];
  jerseyColor: string;
  secondaryColor: string;
  accentColor: string;
  numberColor?: string;
  jerseyStyle: string;
  backgroundId: string;
  showBench: boolean;
  showDetails: boolean;
}