# lib/api-zod/src/generated/types/sharedRosterPlayer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SharedRosterPlayer {
  id: string;
  name: string;
  age: number;
  rating: number;
  nationality: string;
  position: string;
  number: string;
}