# lib/api-zod/src/generated/types/squadGenerationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SquadPlayer } from './squadPlayer';

export interface SquadGenerationResponse {
  players: SquadPlayer[];
}